New version!

+++ P2/Barrientos_Munoz_Vasquez_Practice2/table_data.f90
@@ -4,15 +4,15 @@ module table_mod
 
         contains
         subroutine write_table(table) 
-                REAL*16, intent(out) :: table(:)
+                DOUBLE PRECISION, intent(out) :: table(:)
                 integer :: n, dmn, z, idx
-                real*16 :: beta
+                DOUBLE PRECISION :: beta
 
                 beta = 1.0d0 / temp                
                 dmn = 2.0d0
                 z = dmn*2
                 do n = -2*z, 2*z, 2
-                	idx = (n + 2*z) / 2 + 1
+                        idx = (n + 2*z) / 2 + 1
                         table(idx) = exp(-beta * n)
                 end do
         end subroutine write_table
